refactor: extract EVM scanning logic from support.ts into evm-scanner.ts
